refactor: deprecate the iced_sessionlock and etc
remove the logic to iced_exwlshell, that one contains all of it.
made a application function in iced_exwlshell for easy migration
