+++ iced_exwlshell/src/build_pattern/sessionlock.rs
@@ -1,0 +1,526 @@
+use super::attach;
+use super::daemon::{NameSpace, UpdateFn, with_executor, with_style, with_subscription};
+pub use pattern::application;
+mod pattern {
+    use super::*;
+    use std::borrow::Cow;
+
+    use iced_core::Element;
+    use iced_core::Font;
+    use iced_runtime::Task;
+
+    use crate::actions::ExwlShellCustomActionWithId;
+
+    use crate::DefaultStyle;
+    use crate::settings::LayerShellSettings;
+
+    use crate::Result;
+    use crate::settings::Settings;
+
+    use iced_debug as debug;
+    use iced_program::Program;
+
+    /// The view logic of some [`SingleApplication`].
+    ///
+    /// This trait allows the [`application`] builder to take any closure that
+    /// returns any `Into<Element<'_, Message>>`.
+    pub trait ViewFn<'a, State, Message, Theme, Renderer> {
+        /// Produces the widget of the [`Application`].
+        fn view(&self, state: &'a State) -> Element<'a, Message, Theme, Renderer>;
+    }
+
+    impl<'a, T, State, Message, Theme, Renderer, Widget> ViewFn<'a, State, Message, Theme, Renderer>
+        for T
+    where
+        T: Fn(&'a State) -> Widget,
+        State: 'static,
+        Widget: Into<Element<'a, Message, Theme, Renderer>>,
+    {
+        fn view(&self, state: &'a State) -> Element<'a, Message, Theme, Renderer> {
+            self(state).into()
+        }
+    }
+
+    pub trait BootFn<State, Message> {
+        /// Initializes the [`SingleApplication`] state.
+        fn boot(&self) -> (State, Task<Message>);
+    }
+
+    impl<T, C, State, Message> BootFn<State, Message> for T
+    where
+        T: Fn() -> C,
+        C: IntoBoot<State, Message>,
+    {
+        fn boot(&self) -> (State, Task<Message>) {
+            self().into_boot()
+        }
+    }
+
+    /// The initial state of some [`SingleApplication`].
+    pub trait IntoBoot<State, Message> {
+        /// Turns some type into the initial state of some [`SingleApplication`].
+        fn into_boot(self) -> (State, Task<Message>);
+    }
+
+    impl<State, Message> IntoBoot<State, Message> for State {
+        fn into_boot(self) -> (State, Task<Message>) {
+            (self, Task::none())
+        }
+    }
+
+    impl<State, Message> IntoBoot<State, Message> for (State, Task<Message>) {
+        fn into_boot(self) -> (State, Task<Message>) {
+            self
+        }
+    }
+
+    /// The theme logic of some [`SingleApplication`].
+    ///
+    /// Any implementors of this trait can be provided as an argument to
+    /// [`SingleApplication::theme`].
+    ///
+    /// `iced` provides two implementors:
+    /// - the built-in [`iced::Theme`] itself
+    /// - and any `Fn(&State, iced::window::Id) -> impl Into<Option<iced::Theme>>`.
+    pub trait ThemeFn<State, Theme> {
+        /// Returns the theme of the [`Daemon`] for the current state and window.
+        ///
+        /// If `None` is returned, `iced` will try to use a theme that
+        /// matches the system color scheme.
+        fn theme(&self, state: &State) -> Option<Theme>;
+    }
+
+    impl<State> ThemeFn<State, iced_core::Theme> for iced_core::Theme {
+        fn theme(&self, _state: &State) -> Option<iced_core::Theme> {
+            Some(self.clone())
+        }
+    }
+
+    impl<F, T, State, Theme> ThemeFn<State, Theme> for F
+    where
+        F: Fn(&State) -> T,
+        T: Into<Option<Theme>>,
+    {
+        fn theme(&self, state: &State) -> Option<Theme> {
+            (self)(state).into()
+        }
+    }
+
+    #[derive(Debug)]
+    pub struct SingleApplication<A: Program> {
+        raw: A,
+        settings: Settings,
+        namespace: String,
+    }
+
+    pub fn application<State, Message, Theme, Renderer>(
+        boot: impl BootFn<State, Message>,
+        namespace: impl NameSpace,
+        update: impl UpdateFn<State, Message>,
+        view: impl for<'a> self::ViewFn<'a, State, Message, Theme, Renderer>,
+    ) -> SingleApplication<impl Program<Message = Message, Theme = Theme, State = State>>
+    where
+        State: 'static,
+        Message: 'static
+            + TryInto<ExwlShellCustomActionWithId, Error = Message>
+            + Send
+            + std::fmt::Debug,
+        Theme: DefaultStyle,
+        Renderer: iced_program::Renderer,
+    {
+        use std::marker::PhantomData;
+        struct Instance<State, Message, Theme, Renderer, UpdateFn, ViewFn, BootFn> {
+            update: UpdateFn,
+            view: ViewFn,
+            boot: BootFn,
+            _state: PhantomData<State>,
+            _message: PhantomData<Message>,
+            _theme: PhantomData<Theme>,
+            _renderer: PhantomData<Renderer>,
+        }
+        impl<State, Message, Theme, Renderer, UpdateFn, ViewFn, BootFn> Program
+            for Instance<State, Message, Theme, Renderer, UpdateFn, ViewFn, BootFn>
+        where
+            Message: 'static + TryInto<ExwlShellCustomActionWithId, Error = Message> + Send,
+            Theme: DefaultStyle,
+            Renderer: iced_program::Renderer,
+            UpdateFn: self::UpdateFn<State, Message>,
+            BootFn: self::BootFn<State, Message>,
+            ViewFn: for<'a> self::ViewFn<'a, State, Message, Theme, Renderer>,
+        {
+            type State = State;
+            type Renderer = Renderer;
+            type Message = Message;
+            type Theme = Theme;
+            type Executor = iced_futures::backend::default::Executor;
+
+            fn name() -> &'static str {
+                let name = std::any::type_name::<State>();
+
+                name.split("::").next().unwrap_or("a_cool_application")
+            }
+
+            fn update(
+                &self,
+                state: &mut Self::State,
+                message: Self::Message,
+            ) -> Task<Self::Message> {
+                debug::hot(|| self.update.update(state, message)).into()
+            }
+            fn boot(&self) -> (Self::State, Task<Self::Message>) {
+                self.boot.boot()
+            }
+            fn view<'a>(
+                &self,
+                state: &'a Self::State,
+                _window: iced_core::window::Id,
+            ) -> Element<'a, Self::Message, Self::Theme, Self::Renderer> {
+                debug::hot(|| self.view.view(state))
+            }
+
+            fn settings(&self) -> iced_core::Settings {
+                Default::default()
+            }
+
+            fn window(&self) -> Option<iced_core::window::Settings> {
+                None
+            }
+        }
+        SingleApplication {
+            raw: Instance {
+                update,
+                view,
+                boot,
+                _state: PhantomData,
+                _message: PhantomData,
+                _theme: PhantomData,
+                _renderer: PhantomData,
+            },
+            settings: Settings::default(),
+            namespace: namespace.namespace(),
+        }
+    }
+
+    pub fn with_theme<P: Program>(
+        program: P,
+        f: impl Fn(&P::State) -> Option<P::Theme>,
+    ) -> impl Program<State = P::State, Message = P::Message, Theme = P::Theme> {
+        struct WithTheme<P, F> {
+            program: P,
+            theme: F,
+        }
+
+        impl<P: Program, F> Program for WithTheme<P, F>
+        where
+            F: Fn(&P::State) -> Option<P::Theme>,
+        {
+            type State = P::State;
+            type Message = P::Message;
+            type Theme = P::Theme;
+            type Renderer = P::Renderer;
+            type Executor = P::Executor;
+
+            fn theme(
+                &self,
+                state: &Self::State,
+                _window: iced_core::window::Id,
+            ) -> Option<Self::Theme> {
+                (self.theme)(state)
+            }
+
+            fn name() -> &'static str {
+                P::name()
+            }
+            fn boot(&self) -> (Self::State, Task<Self::Message>) {
+                self.program.boot()
+            }
+
+            fn update(
+                &self,
+                state: &mut Self::State,
+                message: Self::Message,
+            ) -> Task<Self::Message> {
+                self.program.update(state, message)
+            }
+
+            fn view<'a>(
+                &self,
+                state: &'a Self::State,
+                window: iced_core::window::Id,
+            ) -> Element<'a, Self::Message, Self::Theme, Self::Renderer> {
+                self.program.view(state, window)
+            }
+
+            fn subscription(
+                &self,
+                state: &Self::State,
+            ) -> iced_futures::Subscription<Self::Message> {
+                self.program.subscription(state)
+            }
+
+            fn style(&self, state: &Self::State, theme: &Self::Theme) -> crate::Appearance {
+                self.program.style(state, theme)
+            }
+
+            fn scale_factor(&self, state: &Self::State, id: iced_core::window::Id) -> f32 {
+                self.program.scale_factor(state, id)
+            }
+            fn settings(&self) -> iced_core::Settings {
+                Default::default()
+            }
+
+            fn window(&self) -> Option<iced_core::window::Settings> {
+                None
+            }
+        }
+
+        WithTheme { program, theme: f }
+    }
+
+    pub fn with_scale_factor<P: Program>(
+        program: P,
+        f: impl Fn(&P::State) -> f32,
+    ) -> impl Program<State = P::State, Message = P::Message, Theme = P::Theme> {
+        struct WithScaleFactor<P, F> {
+            program: P,
+            scale_factor: F,
+        }
+
+        impl<P: Program, F> Program for WithScaleFactor<P, F>
+        where
+            F: Fn(&P::State) -> f32,
+        {
+            type State = P::State;
+            type Message = P::Message;
+            type Theme = P::Theme;
+            type Renderer = P::Renderer;
+            type Executor = P::Executor;
+
+            fn name() -> &'static str {
+                P::name()
+            }
+
+            fn update(
+                &self,
+                state: &mut Self::State,
+                message: Self::Message,
+            ) -> Task<Self::Message> {
+                self.program.update(state, message)
+            }
+
+            fn view<'a>(
+                &self,
+                state: &'a Self::State,
+                window: iced_core::window::Id,
+            ) -> Element<'a, Self::Message, Self::Theme, Self::Renderer> {
+                self.program.view(state, window)
+            }
+
+            fn subscription(
+                &self,
+                state: &Self::State,
+            ) -> iced_futures::Subscription<Self::Message> {
+                self.program.subscription(state)
+            }
+
+            fn theme(
+                &self,
+                state: &Self::State,
+                window: iced_core::window::Id,
+            ) -> Option<Self::Theme> {
+                self.program.theme(state, window)
+            }
+
+            fn style(&self, state: &Self::State, theme: &Self::Theme) -> crate::Appearance {
+                self.program.style(state, theme)
+            }
+
+            fn scale_factor(&self, state: &Self::State, _id: iced_core::window::Id) -> f32 {
+                (self.scale_factor)(state)
+            }
+            fn boot(&self) -> (Self::State, Task<Self::Message>) {
+                self.program.boot()
+            }
+            fn settings(&self) -> iced_core::Settings {
+                Default::default()
+            }
+
+            fn window(&self) -> Option<iced_core::window::Settings> {
+                None
+            }
+        }
+
+        WithScaleFactor {
+            program,
+            scale_factor: f,
+        }
+    }
+
+    impl<P: Program> SingleApplication<P> {
+        pub fn run(self) -> Result
+        where
+            Self: 'static,
+            P::Message: std::fmt::Debug
+                + Send
+                + 'static
+                + TryInto<ExwlShellCustomActionWithId, Error = P::Message>,
+        {
+            let settings = self.settings;
+
+            #[cfg(all(feature = "debug", not(target_arch = "wasm32")))]
+            let program = {
+                iced_debug::init(iced_debug::Metadata {
+                    name: P::name(),
+                    theme: None,
+                    can_time_travel: cfg!(feature = "time-travel"),
+                });
+
+                super::attach(self.raw)
+            };
+
+            #[cfg(any(not(feature = "debug"), target_arch = "wasm32"))]
+            let program = self.raw;
+
+            #[allow(clippy::needless_update)]
+            let renderer_settings = iced_graphics::Settings {
+                default_font: settings.default_font,
+                default_text_size: settings.default_text_size,
+                antialiasing: if settings.antialiasing {
+                    Some(iced_graphics::Antialiasing::MSAAx4)
+                } else {
+                    None
+                },
+                ..iced_graphics::Settings::default()
+            };
+
+            crate::multi_window::run(
+                program,
+                &self.namespace,
+                settings,
+                renderer_settings,
+                true,
+                None,
+            )
+        }
+
+        pub fn settings(self, settings: Settings) -> Self {
+            Self { settings, ..self }
+        }
+
+        /// Sets the [`Settings::antialiasing`] of the [`SingleApplication`].
+        pub fn antialiasing(self, antialiasing: bool) -> Self {
+            Self {
+                settings: Settings {
+                    antialiasing,
+                    ..self.settings
+                },
+                ..self
+            }
+        }
+
+        /// Sets the default [`Font`] of the [`SingleApplication`].
+        pub fn default_font(self, default_font: Font) -> Self {
+            Self {
+                settings: Settings {
+                    default_font,
+                    ..self.settings
+                },
+                ..self
+            }
+        }
+
+        /// Sets the layershell setting of the [`SingleApplication`]
+        pub fn layer_settings(self, layer_settings: LayerShellSettings) -> Self {
+            Self {
+                settings: Settings {
+                    layer_settings,
+                    ..self.settings
+                },
+                ..self
+            }
+        }
+
+        /// Adds a font to the list of fonts that will be loaded at the start of the [`SingleApplication`].
+        pub fn font(mut self, font: impl Into<Cow<'static, [u8]>>) -> Self {
+            self.settings.fonts.push(font.into());
+            self
+        }
+
+        /// set the default_text_size
+        pub fn default_text_size<Pixels: Into<iced_core::Pixels>>(
+            self,
+            default_text_size: Pixels,
+        ) -> Self {
+            Self {
+                settings: Settings {
+                    default_text_size: default_text_size.into(),
+                    ..self.settings
+                },
+                ..self
+            }
+        }
+
+        pub fn style(
+            self,
+            f: impl Fn(&P::State, &P::Theme) -> crate::Appearance,
+        ) -> SingleApplication<impl Program<State = P::State, Message = P::Message, Theme = P::Theme>>
+        {
+            SingleApplication {
+                raw: with_style(self.raw, move |state, theme| f(state, theme)),
+                settings: self.settings,
+                namespace: self.namespace,
+            }
+        }
+        /// Sets the subscription logic of the [`SingleApplication`].
+        pub fn subscription(
+            self,
+            f: impl Fn(&P::State) -> iced_futures::Subscription<P::Message>,
+        ) -> SingleApplication<impl Program<State = P::State, Message = P::Message, Theme = P::Theme>>
+        {
+            SingleApplication {
+                raw: with_subscription(self.raw, move |state| debug::hot(|| f(state))),
+                settings: self.settings,
+                namespace: self.namespace,
+            }
+        }
+
+        /// Sets the theme logic of the [`SingleApplication`].
+        pub fn theme(
+            self,
+            f: impl ThemeFn<P::State, P::Theme>,
+        ) -> SingleApplication<impl Program<State = P::State, Message = P::Message, Theme = P::Theme>>
+        {
+            SingleApplication {
+                raw: with_theme(self.raw, move |state| debug::hot(|| f.theme(state))),
+                settings: self.settings,
+                namespace: self.namespace,
+            }
+        }
+
+        /// Sets the scale factor of the [`SingleApplication`].
+        pub fn scale_factor(
+            self,
+            f: impl Fn(&P::State) -> f32,
+        ) -> SingleApplication<impl Program<State = P::State, Message = P::Message, Theme = P::Theme>>
+        {
+            SingleApplication {
+                raw: with_scale_factor(self.raw, move |state| debug::hot(|| f(state))),
+                settings: self.settings,
+                namespace: self.namespace,
+            }
+        }
+        /// Sets the executor of the [`SingleApplication`].
+        pub fn executor<E>(
+            self,
+        ) -> SingleApplication<impl Program<State = P::State, Message = P::Message, Theme = P::Theme>>
+        where
+            E: iced_futures::Executor,
+        {
+            SingleApplication {
+                raw: with_executor::<P, E>(self.raw),
+                settings: self.settings,
+                namespace: self.namespace,
+            }
+        }
+    }
+}
