+++ iced_exwlshell/src/lib.rs
@@ -39,7 +39,7 @@ pub mod reexport {
 mod ime_preedit;
 
 #[cfg(feature = "macros")]
-pub use iced_exwlshell_macros::to_exwlshell_message;
+pub use iced_exwlshell_macros::{to_exwlshell_message, to_sessionlock_message};
 
 pub use error::Error;
 pub use iced_wayland_subscriber::shell;
@@ -61,5 +61,6 @@ use iced_core::theme::Base as DefaultStyle;
 
 // layershell application
 pub use build_pattern::daemon;
+pub use build_pattern::sessionlock;
 
 pub use settings::Settings;
