--- iced_sessionlock/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "iced_sessionlock"
-authors.workspace = true
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-keywords.workspace = true
-repository.workspace = true
-readme = "README.md"
-description = "sessionlock binding for iced"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-default = ["linux-theme-detection", "macros"]
-
-macros = ["dep:iced_sessionlock_macros"]
-debug = ["iced_debug/enable"]
-time-travel = ["iced_exdevtools/time-travel", "debug"]
-unconditional-rendering = []
-linux-theme-detection = ["dep:mundy"]
-
-[dependencies]
-iced_sessionlock_macros = { workspace = true, optional = true }
-iced_renderer.workspace = true
-iced_runtime.workspace = true
-iced_core.workspace = true
-iced_futures.workspace = true
-iced_graphics.workspace = true
-iced_program.workspace = true
-iced_debug.workspace = true
-iced_exdevtools.workspace = true
-tracing.workspace = true
-thiserror.workspace = true
-sessionlockev.workspace = true
-iced_wayland_subscriber.workspace = true
-futures.workspace = true
-window_clipboard.workspace = true
-log.workspace = true
-
-# workspace features cannot be optional so we need to define the dependency here
-mundy = { version = "0.2.0", optional = true }
-
-[dev-dependencies]
-iced.workspace = true
