--- iced_sessionlock/src/actions.rs
@@ -1,12 +1,0 @@
-use iced_core::mouse::Interaction;
-use sessionlockev::id::Id as SessionId;
-#[allow(unused)]
-#[derive(Debug, Clone)]
-pub(crate) enum SessionShellAction {
-    Mouse(Interaction),
-    RedrawAll,
-    RedrawWindow(SessionId),
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct UnLockAction;
