--- iced_sessionlock/src/clipboard.rs
@@ -1,87 +1,0 @@
-use std::sync::atomic::{AtomicBool, Ordering};
-
-use iced_core::Clipboard;
-use iced_core::clipboard::Kind;
-use sessionlockev::WindowWrapper;
-
-static DISABLED: AtomicBool = AtomicBool::new(false);
-
-pub(crate) fn set_disabled() {
-    DISABLED.store(true, Ordering::Relaxed);
-}
-
-pub(crate) fn is_disabled() -> bool {
-    DISABLED.load(Ordering::Relaxed)
-}
-
-pub struct SessionLockClipboard {
-    state: State,
-}
-
-enum State {
-    Connected(window_clipboard::Clipboard),
-    Unavailable,
-}
-
-impl SessionLockClipboard {
-    /// Creates a new [`Clipboard`] for the given window.
-    pub fn connect(window: &WindowWrapper) -> Self {
-        #[allow(unsafe_code)]
-        let state = unsafe { window_clipboard::Clipboard::connect(window) }
-            .ok()
-            .map(State::Connected)
-            .unwrap_or(State::Unavailable);
-
-        Self { state }
-    }
-
-    /// Creates a new [`Clipboard`] that isn't associated with a window.
-    /// This clipboard will never contain a copied value.
-    #[allow(unused)]
-    pub fn unconnected() -> Self {
-        Self {
-            state: State::Unavailable,
-        }
-    }
-
-    /// Reads the current content of the [`Clipboard`] as text.
-    pub fn read(&self, kind: Kind) -> Option<String> {
-        match &self.state {
-            State::Connected(clipboard) => match kind {
-                Kind::Standard => clipboard.read().ok(),
-                Kind::Primary => clipboard.read_primary().and_then(Result::ok),
-            },
-            State::Unavailable => None,
-        }
-    }
-
-    /// Writes the given text contents to the [`Clipboard`].
-    pub fn write(&mut self, kind: Kind, contents: String) {
-        match &mut self.state {
-            State::Connected(clipboard) => {
-                let result = match kind {
-                    Kind::Standard => clipboard.write(contents),
-                    Kind::Primary => clipboard.write_primary(contents).unwrap_or(Ok(())),
-                };
-
-                match result {
-                    Ok(()) => {}
-                    Err(error) => {
-                        log::warn!("error writing to clipboard: {error}");
-                    }
-                }
-            }
-            State::Unavailable => {}
-        }
-    }
-}
-
-impl Clipboard for SessionLockClipboard {
-    fn read(&self, kind: Kind) -> Option<String> {
-        self.read(kind)
-    }
-
-    fn write(&mut self, kind: Kind, contents: String) {
-        self.write(kind, contents);
-    }
-}
