--- iced_sessionlock/src/conversion/keymap.rs
@@ -1,1 +1,0 @@
-pub use iced_exdevtools::keymap::*;
