--- iced_sessionlock/src/error.rs
@@ -1,27 +1,0 @@
-use iced_futures::futures;
-use sessionlockev::SessionLockEventError;
-
-/// An error that occurred while running an application.
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    /// The futures executor could not be created.
-    #[error("the futures executor could not be created")]
-    ExecutorCreationFailed(futures::io::Error),
-
-    /// The application window could not be created.
-    #[error("the application window could not be created")]
-    WindowCreationFailed(Box<dyn std::error::Error + Send + Sync>),
-
-    /// The application graphics context could not be created.
-    #[error("the application graphics context could not be created")]
-    GraphicsCreationFailed(iced_graphics::Error),
-
-    #[error("Error during dispatch")]
-    WaylandDispatchFailed(#[from] SessionLockEventError),
-}
-
-impl From<iced_graphics::Error> for Error {
-    fn from(error: iced_graphics::Error) -> Error {
-        Error::GraphicsCreationFailed(error)
-    }
-}
