--- iced_sessionlock/src/settings.rs
@@ -1,58 +1,0 @@
-use iced_core::{Font, Pixels};
-use sessionlockev::reexport::wayland_client::Connection;
-use std::borrow::Cow;
-
-#[derive(Debug)]
-pub struct Settings {
-    /// The identifier of the application.
-    ///
-    /// If provided, this identifier may be used to identify the application or
-    /// communicate with it through the windowing system.
-    pub id: Option<String>,
-
-    /// The data needed to initialize an Application.
-    ///
-    /// The fonts to load on boot.
-    pub fonts: Vec<Cow<'static, [u8]>>,
-
-    /// The default [`Font`] to be used.
-    ///
-    /// By default, it uses [`Family::SansSerif`](iced::font::Family::SansSerif).
-    pub default_font: Font,
-
-    /// The text size that will be used by default.
-    ///
-    /// The default value is `16.0`.
-    pub default_text_size: Pixels,
-
-    /// If set to true, the renderer will try to perform antialiasing for some
-    /// primitives.
-    ///
-    /// Enabling it can produce a smoother result in some widgets, like the
-    /// `Canvas`, at a performance cost.
-    ///
-    /// By default, it is disabled.
-    ///
-    pub antialiasing: bool,
-    /// set the used wayland connection, all wayland object will share it, and they can be used by
-    /// each other.
-    pub with_connection: Option<Connection>,
-
-    /// Where the runtime announces the lock surfaces it creates and the
-    /// monitors they cover. Pair it with
-    /// [`iced_wayland_subscriber::shell::channel`] and keep the receiver.
-    pub shell_broadcast: iced_wayland_subscriber::shell::ShellSender,
-}
-impl Default for Settings {
-    fn default() -> Self {
-        Settings {
-            id: None,
-            fonts: Vec::new(),
-            default_font: Font::default(),
-            default_text_size: Pixels(16.0),
-            antialiasing: false,
-            with_connection: None,
-            shell_broadcast: iced_wayland_subscriber::shell::channel().0,
-        }
-    }
-}
