--- iced_sessionlock/tests/macro_test.rs
@@ -1,15 +1,0 @@
-#![cfg(feature = "macros")]
-
-use iced_sessionlock::to_session_message;
-
-#[test]
-fn test_macro() {
-    #[allow(dead_code)]
-    #[to_session_message]
-    #[derive(Debug, Clone)]
-    enum TestEnum {
-        TestA,
-    }
-    let e = TestEnum::UnLock;
-    let _ = e.clone();
-}
