--- iced_sessionlock_macros/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "iced_sessionlock_macros"
-authors.workspace = true
-edition.workspace = true
-version.workspace = true
-license.workspace = true
-repository.workspace = true
-description.workspace = true
-keywords.workspace = true
-readme.workspace = true
-
-[lib]
-proc-macro = true
-
-[features]
-default = []
-# Only nightly
-diagnostics = ["darling/diagnostics"]
-
-[dependencies]
-darling.workspace = true
-manyhow.workspace = true
-proc-macro2.workspace = true
-quote.workspace = true
-syn.workspace = true
